fix type of n option parameter.

+++ head.go
@@ -8,10 +8,10 @@ import (
 )
 
 // Head display first line of a file
-func Head(file io.Reader, n int, w io.Writer) error {
+func Head(file io.Reader, n uint64, w io.Writer) error {
 	buf := new(bytes.Buffer)
 	reader := bufio.NewReader(file)
-	for i := 0; i < n; i++ {
+	for i := uint64(0); i < n; i++ {
 		line, err := reader.ReadBytes('\n')
 		if err != nil && err != io.EOF {
 			return err
